requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==3.2.25
django-dotenv==1.4.2
django-heroku==0.3.1
django-storages==1.14.2
djangorestframework==3.15.1
executing==2.0.1
gunicorn==21.2.0
packaging==23.2
pexpect==4.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
sqlparse==0.4.4
stack-data==0.6.3
typing_extensions==4.10.0
whitenoise==6.6.0
python-dotenv==1.0.1
load_dotenv==0.1.0
django_heroku==0.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
